+++ ui.lua
@@ -129,4 +129,91 @@ function Menu:mousereleased(x, y, button)
 end
 
 
+-- Dialogue class.
+-- TODO: Add sound effect when clicking dialogue box.
+
+local Dialogue = {}
+Dialogue.__index = Dialogue
+ui.Dialogue = Dialogue
+
+function Dialogue.new(lines)
+    return setmetatable({
+        -- These are fixed.
+        x = 20, y = 360, w = 600, h = 100,
+        lines = lines,
+        speed = 20,
+
+        -- These are updated.
+        current   = 1,
+        revealed  = 0,
+        displayed = "",
+        elapsed   = 0,
+        finished  = false,
+    }, Dialogue)
+end
+
+function Dialogue:update(dt)
+    if self.finished then
+        return
+    end
+
+    local text       = self.lines[self.current + 1]
+    local textLength = #text
+    self.elapsed     = self.elapsed + dt * self.speed
+
+    -- While loop lets multiple characters be consumed if dt large.
+    while self.elapsed >= 1 and self.revealed <= textLength do
+        local ch       = text:sub(self.revealed, self.revealed)
+        self.displayed = self.displayed .. ch
+        self.revealed  = self.revealed + 1
+        self.elapsed   = self.elapsed - 1
+    end
+    if self.revealed > textLength then
+        self.finished = true
+    end
+end
+
+function Dialogue:draw()
+    -- Box for speaker name.
+    love.graphics.rectangle("fill", self.x, self.y - game.fontHeight - 5, 60, game.fontHeight)
+    -- Box for the text.
+    love.graphics.rectangle("fill", self.x, self.y, self.w, self.h)
+
+    love.graphics.setColor(0, 0, 0)
+    local n = self.lines[self.current]
+    local name = nil
+    if n == 1 then
+        name = "Ada"
+    elseif n == 2 then
+        name = "You"
+    else
+        name = "..."
+    end
+
+    love.graphics.print(name, self.x, self.y - game.fontHeight - 5)
+    love.graphics.printf(self.displayed, self.x, self.y, self.w)
+end
+
+function Dialogue:mousepressed(mx, my, button)
+    if mx >= self.x and
+        mx <= self.x + self.w and
+        my >= self.y and
+        my <= self.y + self.h then
+
+        if self.finished then
+            if self.current < #self.lines - 1 then
+                self.current   = self.current + 2
+                self.revealed  = 0
+                self.displayed = ""
+                self.elapsed   = 0
+                self.finished  = false
+            end
+        else
+            self.displayed = self.lines[self.current + 1]
+            self.finished = true
+        end
+    end
+end
+
+
 return ui