+++ CHANGELOG.md
@@ -11,3 +11,6 @@ All notable changes to this project will be documented in this file.
 - 2 characters: Ichinose Kotomi and Kanbe Kotori.
 - Pattern pngs.
 - Audio for home screen.
+- Ep1: Hajimari.
+- Dialogue class.
+- Typing effect in dialogue box.